fix: delete user from db and keycloak, change balance field

+++ src/main/java/group/aist/cinema/dto/common/BalanceDTO.java
@@ -15,6 +15,6 @@ public class BalanceDTO {
 
     private String currency;
 
-    private Double amount;
+    private BigDecimal amount;
 
 }
